Updated constants and Added Feedback Devices to Talons
Updated the constants to have default values of
one which need to be collected the next time.

The talons now have a Quadrature Encoder
specified as their feedback device which allows
them to figure out how far the robot has mode.

The Robot State appears to be working with the known
limitation of now being able to turn too fast or else
the gyroscope gets confused.